corrected generation of json file

+++ src_toml_interface/from_toml.cc
@@ -132,7 +132,7 @@ namespace GC_namespace {
     } else if ( node.is_array_of_tables() ) {
 
       gc.set_vector();
-      size_t i{0};
+      unsigned i{0};
       for ( const auto & value : *node.as_array() ) {
         ok = TOML_to_GC( value, gc[i++] );
         if ( !ok ) break;
@@ -172,7 +172,7 @@ namespace GC_namespace {
       } else {
 
         gc.set_vector();
-        size_t i{0};
+        unsigned i{0};
         for ( const auto& value : *ary ) {
           ok = TOML_to_GC( value, gc[i++] );
           if ( !ok ) break;
